+++ services/follower-api/src/main/java/com/twitter/controller/FollowController.java
@@ -1,12 +1,18 @@
 package com.twitter.controller;
 
 import com.twitter.common.aspect.LoggableRequest;
+import com.twitter.dto.filter.FollowerFilter;
 import com.twitter.dto.request.FollowRequestDto;
 import com.twitter.dto.response.FollowResponseDto;
+import com.twitter.dto.response.FollowerResponseDto;
 import com.twitter.service.FollowService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
+import org.springframework.data.web.PageableDefault;
+import org.springframework.data.web.PagedModel;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -44,10 +50,24 @@ public class FollowController implements FollowApi {
     @LoggableRequest
     @DeleteMapping("/{followerId}/{followingId}")
     @Override
-    public ResponseEntity<Void> deleteFollow(@PathVariable("followerId") UUID followerId,
-                                             @PathVariable("followingId") UUID followingId) {
+    public ResponseEntity<Void> deleteFollow(
+        @PathVariable("followerId") UUID followerId,
+        @PathVariable("followingId") UUID followingId) {
         followService.unfollow(followerId, followingId);
         return ResponseEntity.noContent().build();
+    }
+
+    /**
+     * @see FollowApi#getFollowers
+     */
+    @LoggableRequest
+    @GetMapping("/{userId}/followers")
+    @Override
+    public PagedModel<FollowerResponseDto> getFollowers(
+        @PathVariable("userId") UUID userId,
+        @ModelAttribute FollowerFilter filter,
+        @PageableDefault(size = 10, sort = "createdAt", direction = Sort.Direction.DESC) Pageable pageable) {
+        return followService.getFollowers(userId, filter, pageable);
     }
 }